move accessors to own file
