+++ pkg/reconciler/labeler/labels.go
@@ -31,14 +31,6 @@ import (
 	v1 "knative.dev/serving/pkg/apis/serving/v1"
 )
 
-// accessor defines an abstraction for manipulating labeled entity
-// (Configuration, Revision) with shared logic.
-type accessor interface {
-	get(ns, name string) (kmeta.Accessor, error)
-	list(ns, name string) ([]kmeta.Accessor, error)
-	patch(ns, name string, pt types.PatchType, p []byte) error
-}
-
 // syncLabels makes sure that the revisions and configurations referenced from
 // a Route are labeled with route labels.
 func (c *Reconciler) syncLabels(ctx context.Context, r *v1.Route) error {
@@ -204,45 +196,4 @@ func (r *revision) list(ns, name string) ([]kmeta.Accessor, error) {
 		kl = append(kl, r)
 	}
 	return kl, err
-}
-
-// patch implements accessor
-func (r *revision) patch(ns, name string, pt types.PatchType, p []byte) error {
-	_, err := r.r.client.ServingV1().Revisions(ns).Patch(name, pt, p)
-	return err
-}
-
-// configuration is an implementation of accessor for Configurations
-type configuration struct {
-	r *Reconciler
-}
-
-// configuration implements accessor
-var _ accessor = (*configuration)(nil)
-
-// get implements accessor
-func (c *configuration) get(ns, name string) (kmeta.Accessor, error) {
-	return c.r.configurationLister.Configurations(ns).Get(name)
-}
-
-// list implements accessor
-func (c *configuration) list(ns, name string) ([]kmeta.Accessor, error) {
-	rl, err := c.r.configurationLister.Configurations(ns).List(labels.SelectorFromSet(labels.Set{
-		serving.RouteLabelKey: name,
-	}))
-	if err != nil {
-		return nil, err
-	}
-	// Need a copy to change types in Go
-	kl := make([]kmeta.Accessor, 0, len(rl))
-	for _, r := range rl {
-		kl = append(kl, r)
-	}
-	return kl, err
-}
-
-// patch implements accessor
-func (c *configuration) patch(ns, name string, pt types.PatchType, p []byte) error {
-	_, err := c.r.client.ServingV1().Configurations(ns).Patch(name, pt, p)
-	return err
 }
